server/requirements.txt
absl-py==2.1.0
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.13
decorator==5.2.1
dnspython==2.7.0
easydict==1.13
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
faiss-gpu==1.7.2
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
insightface==0.7.3
ipykernel==6.29.5
ipython==8.34.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mysql-connector-python==9.2.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu12==12.5.1.3
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.23.4
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.6.77
onnx==1.17.0
onnxruntime-gpu==1.21.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optional-django==0.3.0
optree==0.14.1
orjson==3.10.15
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickle-mixin==1.0.2
pillow==11.1.0
platformdirs==4.3.6
prettytable==3.15.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.2.1
react==4.3.0
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
shapely==2.0.7
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
stringzilla==3.12.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
